Redux added to submission form
