+++ src/components/forms/SubmissionForm/SubmissionForm.css
@@ -150,3 +150,9 @@ textarea {
   color: #e63946; 
 }
 
+.error-message {
+  padding-top: 5px;
+  color: #e63946;
+  font-size: 14px;
+}
+
