+++ src/redux/actions/submissionFormActions.js
@@ -1,0 +1,6 @@
+export const submitFormData = (formData) => {
+    return {
+      type: SUBMIT_FORM_DATA,
+      payload: formData,
+    };
+  };
