+++ src/redux/index.js
@@ -1,0 +1,1 @@
+export const SUBMIT_FORM_DATA = "SUBMIT_FORM_DATA";
